build: fix warnings before enabling -Wall
A bunch of warnings creaped up as the -Wall and other warning
and error flags were added to the build. Fix these.

This also introduces TPM2B_TYPE_INIT() macro, which can be
used to get rid of the crazy sizeof(x) - 2 logic that was
prevelent.

+++ lib/pcr.c
@@ -38,7 +38,7 @@ static int pcr_parse_selection(const char *str, int len, TPMS_PCR_SELECTION *pcr
         return -1;
     }
 
-    if (strLeft - str > sizeof(buf) - 1) {
+    if ((size_t)(strLeft - str) > sizeof(buf) - 1) {
         return -1;
     }
 
@@ -123,7 +123,7 @@ int pcr_parse_list(const char *str, int len, TPMS_PCR_SELECTION *pcrSel) {
             len = 0;
         }
 
-        if (lenCurrent > sizeof(buf) - 1) {
+        if ((size_t)lenCurrent > sizeof(buf) - 1) {
             return -1;
         }
